Modification of the signatures for the interface functions.
It appears that there is no need to set errno in case there is an error to be raised.
Therefore, better not to do so for performance reasons.
Moreover, this change overrides a little mistake in the use of returned values.
These were still supposed to be of type int in simplifier.c because of a previous incomplete change.